requirements.txt
flask==2.3.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gunicorn==21.2.0
groq==0.4.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
supabase==2.0.2
weasyprint==60.2
Werkzeug==2.3.7
